feat: refactor error gathering to use compiler's gather_errors function and remove unused classes
